\name{as.data.frame.bond}
\Rdversion{1.1}
\alias{as.data.frame.bond}
\alias{as.data.frame.portfolio.bond}
\title{
Returns a data-frame summary of object.
}
\description{
Returns a data-frame summary of bond or portfolio.
}
\usage{
	\method{as.data.frame}{bond}(x, ...)
}
\arguments{
   \item{x}{
	Bond or portfolio.bond object
   }
   \item{\dots}{ Pass-alongs. }
}
