\name{history.market}
\alias{history.market}
\title{
Creates a history.market object
}
\description{
A history.market object stores markets over time
}
\usage{
history.market(mkts, ...)
}
\arguments{
  \item{mkts}{
A list of market objects
  }
  \item{\dots}{
Pass-alongs
  }
}

\value{
a history.market object
}

\seealso{
For automated market history creation, see \code{\link{genHistory.market}}, ~~~
}
