\name{print.market}
\Rdversion{1.1}
\alias{print.market}
\alias{print.sum.market.bond}
\alias{print.bond}
\alias{print.sum.bond}
\alias{print.cash}
\alias{print.account}
\alias{print.sum.account}
\alias{print.portfolio.bond}
\alias{print.sum.portfolio.bond}
\alias{print.history.market}
\alias{print.market.bond}
\title{
Display characteristics of maRketSim object on console
}
\description{
Prints the objects on the screen
}
\usage{
	\method{print}{market}(x,...)
	\method{print}{sum.market.bond}(x,...)
	\method{print}{bond}(x,...)
	\method{print}{sum.bond}(x,...)
	\method{print}{cash}(x,...)
	\method{print}{account}(x,...)
	\method{print}{sum.account}(x,...)
	\method{print}{portfolio.bond}(x,...)
	\method{print}{sum.portfolio.bond}(x,...)
	\method{print}{history.market}(x,...)
	\method{print}{market.bond}(x,...)
}
\arguments{
	\item{x}{
	A maRketSim object
	}
	\item{\dots}{ Pass-alongs. }
}
