\name{summary.vasicek.discrete}
\alias{summary.vasicek.discrete}
\title{
Summarize a Vasicek model
}
\description{
Summarize a Vasicek model
}
\usage{
	\method{summary}{vasicek.discrete}(object, lambda, ...)
}
\arguments{
  \item{object}{
vasicek.discrete object
}
  \item{lambda}{
Lambda is a measure of excess return per unit of standard deviation (risk-reward relationship), and is constant across all terms
}
  \item{\dots}{
Pass-alongs
}
}
\value{
sum.vasicek.discrete object
}
